# openapi.yaml
openapi: 3.0.1
info:
  title: REST API documentation
  description: >-
    Приложение по <a href='https://javaops.ru/view/topjava2'>курсу TopJava-2</a>
    (решение выпускного проекта)<br>Для /api/admin использовать логин / пароль:
    admin@gmail.com / admin<br>Для /api/profile использовать логин / пароль:
    user@yandex.ru / password
  contact:
    name: Konstantin Shurukhin
    email: kshurukhin@gmail.com
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - basicAuth: []
paths:
...
  /api/votes:
    get:
      tags:
        - vote-controller
      operationId: getVotes
      parameters:
        - name: restaurantId
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: userId
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VoteTo'
...
components:
  schemas:
...
    User:
      required:
        - email
        - name
        - password
        - registered
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          maxLength: 100
          minLength: 2
          type: string
        email:
          maxLength: 100
          minLength: 0
          type: string
        password:
          maxLength: 100
          minLength: 5
          type: string
          writeOnly: true
        enabled:
          type: boolean
        registered:
          type: string
          format: date-time
          readOnly: true
        roles:
          uniqueItems: true
          type: array
          items:
            type: string
            enum:
              - USER
              - ADMIN
    Dish:
      required:
        - name
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 2
          type: string
        price:
          type: number
          format: double
...
    RestaurantTo:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          maxLength: 100
          minLength: 2
          type: string
        rating:
          type: integer
          format: int32
        menu:
          type: array
          items:
            $ref: '#/components/schemas/Dish'
    VoteTo:
      type: object
      properties:
        id:
          type: integer
          format: int32
        date:
          type: string
          format: date
        user:
          $ref: '#/components/schemas/User'
        restaurant:
          $ref: '#/components/schemas/RestaurantTo'
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic